be/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Cors==6.0.2
Flask-Limiter==4.1.1
psycopg2-binary==2.9.11
marshmallow==4.2.1
bcrypt==5.0.0
python-dotenv==1.2.1
cloudinary==1.44.1
requests==2.32.5
gunicorn==25.0.1
waitress==3.0.2
